+++ src/Messenger.Client.Extensions/Properties/AssemblyInfo.cs
@@ -18,4 +18,4 @@ using System.Runtime.InteropServices;
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("da848ca4-f166-4151-aafd-29864365d454")]
 
-[assembly: AssemblyVersion("0.3.1")]
+[assembly: AssemblyVersion("0.4.2")]