+++ src/Messenger.Client/Objects/MessengerDomainActionType.cs
@@ -1,0 +1,9 @@
+﻿namespace Messenger.Client.Objects
+{
+    public static class MessengerDomainActionType
+    {
+        public const string Add = "add";
+
+        public const string Remove = "remove";
+    }
+}