+++ src/Messenger.Client/Objects/MessengerGreeting.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Messenger.Client.Objects
+{
+    public class MessengerGreeting
+    {
+        public string Text { get; set; }
+    }
+}