+++ src/Messenger.Client/Objects/MessengerGreetingTextSetting.cs
@@ -1,0 +1,9 @@
+﻿namespace Messenger.Client.Objects
+{
+    public class MessengerGreetingTextSetting : MessengerThreadSetting
+    {
+        public override string SettingType => "greeting";
+
+        public MessengerGreeting Greeting { get; set; }
+    }
+}