+++ src/Messenger.Client/Objects/MessengerPayloadCallToAction.cs
@@ -1,0 +1,7 @@
+﻿namespace Messenger.Client.Objects
+{
+    public class MessengerPayloadCallToAction
+    {
+        public string Payload { get; set; }
+    }
+}