+++ src/Messenger.Client/Objects/MessengerTemplateType.cs
@@ -1,0 +1,22 @@
+﻿
+namespace Messenger.Client.Objects
+{
+    public static class MessengerTemplateType
+    {
+        public const string AirlineBoardingPass = "airline_boardingpass";
+
+        public const string AirlineCheckin = "airline_checkin";
+
+        public const string AirlineItiniery = "airline_itinerary";
+
+        public const string AirlineFlightUpdate = "airline_update";
+
+        public const string Button = "button";
+
+        public const string Generic = "generic";
+
+        public const string List = "list";
+
+        public const string Receipt = "receipt";
+    }
+}