+++ src/Messenger.Client/Objects/MessengerTopElementStyles.cs
@@ -1,0 +1,9 @@
+﻿namespace Messenger.Client.Objects
+{
+    public static class MessengerTopElementStyles
+    {
+        public const string Compact = "compact";
+
+        public const string Large = "large";
+    }
+}