+++ src/Messenger.Client/Services/IMessengerThreadSettingsService.cs
@@ -1,10 +1,13 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
 using Messenger.Client.Objects;
 
 namespace Messenger.Client.Services
 {
     public interface IMessengerThreadSettingsService
     {
-        Task<MessengerResponse<string>>  Save(MessengerThreadSetting setting, string accessToken);
+        Task<MessengerResponse<string>> Save(MessengerThreadSetting setting, string accessToken);
+
+        Task<MessengerResponse<string>> WhitelistDomains(IEnumerable<string> domains, string accessToken);
     }
 }